refactor get chain icon src to fxn and hook
